Break up slow tests
